+++ .changeset/selfish-sheep-call.md
@@ -1,0 +1,5 @@
+---
+'@finos/legend-dev-utils': minor
+---
+
+Add test mocks, polyfills and test setup scripts for Jest.
