+++ package.json
@@ -17,7 +17,7 @@
     "check:env-info": "node ./scripts/workflow/checkEnvInfo.js",
     "check:format": "prettier --list-different \"(scripts|docs)/**/*.{md,json,mjs,cjs,js,ts,tsx,html,scss,css}\" \"packages/*/**/*.{md,json,mjs,cjs,js,ts,tsx,html,scss,css}\" && yarn sort-package-json --check \"package.json\" \"packages/*/package.json\"",
     "check:pkg-constraints": "yarn constraints && echo \"Checking for circular dependencies...\" && yarn workspaces foreach --topological-dev exec \"echo '' >/dev/null\"",
-    "check:project-ref": "node ./scripts/typescript/checkProjectReferenceConfigs.js",
+    "check:project-ref": "node ./scripts/workflow/checkProjectReferenceConfigs.js",
     "check:tsc": "yarn build:tsc",
     "check:update": "yarn upgrade-interactive",
     "clean": "rimraf \"build\" && yarn workspaces foreach --all --topological-dev --parallel --verbose run clean",
@@ -79,7 +79,6 @@
     "@finos/eslint-plugin-legend-studio": "workspace:*",
     "@finos/legend-dev-utils": "workspace:*",
     "@finos/stylelint-config-legend-studio": "workspace:*",
-    "@juggle/resize-observer": "3.3.1",
     "@types/jest": "27.0.2",
     "@types/node": "16.11.6",
     "chalk": "4.1.2",
