+++ packages/legend-dev-utils/jest.config.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-import base from '../../scripts/jest/jest.config.base.js';
+import base from '../../scripts/test/jest.config.base.js';
 import { loadJSON } from './DevUtils.js';
 
 const packageJson = loadJSON('./package.json');
