+++ packages/legend-dev-utils/jest/disallowConsoleError.js
