+++ packages/legend-dev-utils/jest/setupTestEnvironment.js
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-// increase timeout for some long running tests
+// Increase timeout for some long running tests
 jest.setTimeout(10000);
 // Runs failed tests n-times until they pass or until the max number of retries is exhausted.
 jest.retryTimes(2);
