+++ packages/legend-dev-utils/package.json
@@ -21,7 +21,12 @@
   "sideEffects": false,
   "type": "module",
   "exports": {
+    "./jest/setupTestEnvironment": "./jest/setupTestEnvironment.js",
     "./jest/setupJestExpectExtension": "./jest/setupJestExpectExtension.js",
+    "./jest/disallowConsoleError": "./jest/disallowConsoleError.js",
+    "./jest/blockFetch": "./jest/blockFetch.js",
+    "./jest/handleUnhandledRejection": "./jest/handleUnhandledRejection.js",
+    "./jest/setupDOMPolyfills": "./jest/setupDOMPolyfills.js",
     "./mocks/fileMock": "./mocks/fileMock.js",
     "./ChangesetUtils": "./ChangesetUtils.js",
     "./ChangesetChangelogUtils": "./ChangesetChangelogUtils.cjs",
@@ -52,6 +57,7 @@
     "@changesets/logger": "0.0.5",
     "@changesets/read": "0.5.0",
     "@changesets/write": "0.1.5",
+    "@juggle/resize-observer": "3.3.1",
     "@manypkg/get-packages": "1.1.2",
     "@pmmmwh/react-refresh-webpack-plugin": "0.5.1",
     "autoprefixer": "10.4.0",
