+++ scripts/workflow/checkProjectReferenceConfigs.js
